Rename package from com.reactapp to com.retina
