+++ android/app/build.gradle
@@ -77,9 +77,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.reactapp"
+    namespace "com.retina"
     defaultConfig {
-        applicationId "com.reactapp"
+        applicationId "com.retina"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
