+++ android/app/src/main/java/com/retina/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.reactapp
+package com.retina
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
